moved sections to files; set up template for projects; changed sun colour back to black; menu button display on mobile responsiveness (no functionality yet)
