+++ src/App.css
@@ -8,6 +8,27 @@
 }
 
 /* fonts */
+.onest-lightest {
+  font-family: "Onest", sans-serif;
+  font-optical-sizing: auto;
+  font-weight: 100;
+  font-style: normal;
+}
+
+.onest-lighter {
+  font-family: "Onest", sans-serif;
+  font-optical-sizing: auto;
+  font-weight: 200;
+  font-style: normal;
+}
+
+.onest-light {
+  font-family: "Onest", sans-serif;
+  font-optical-sizing: auto;
+  font-weight: 300;
+  font-style: normal;
+}
+
 .onest {
   font-family: "Onest", sans-serif;
   font-optical-sizing: auto;
