+++ src/App.tsx
@@ -2,11 +2,13 @@ import './App.css'
 import { useState } from 'react';
 import Nav from './components/navbar/Nav';
 import { DarkModeContext } from './components/navbar/DarkModeContext';
-import Footer from './components/main/Footer';
+import Footer from './components/sections/Footer';
 import filledheart from './assets/filledheart.svg';
 import heart from './assets/pinkheart.svg';
 import GradientTypewriteText from './components/main/GradientTypewriteText';
-import Card from './components/Card';
+import Card from './components/main/Card';
+import Intro from './components/sections/Intro';
+import Projects from './components/sections/Projects';
 
 function App() {
   const [darkMode, setDarkMode] = useState<string>(localStorage.getItem('dark-mode')!);
@@ -27,68 +29,8 @@ function App() {
           </section>
 
           <section id="content" className='pt-10 pl-8 flex flex-col'>
-            <div className='onest text-2xl inline-block'>
-              I&nbsp;
-              <p className='bg-clip-text text-transparent inline-block bg-gradient-to-r from-pink-500 to-violet-500'>
-                {darkMode ?
-                  <>
-                    l
-                    <img alt='pulsing heart' src={heart} width={18} className='m-[0.1rem] motion-preset-pulse-sm inline-block hover:motion-paused' />
-                    ve&nbsp;
-                  </>
-                  :
-                  <p>
-                    l
-                    <img alt='pulsing heart' src={heart} width={18} className='m-[0.1rem] motion-preset-pulse-sm inline-block hover:motion-paused' />
-                    ve&nbsp;
-                  </p>
-                  // <img alt='pulsing heart' src={filledheart} width={50} className='motion-preset-pulse-sm inline-block hover:motion-paused' />
-                }
-
-              </p>
-              all
-              <p className='motion-rotate-loop-2 motion-duration-[4000ms] motion-delay-2000 hover:motion-paused motion-preset-oscillate motion-preset-float-sm bg-gradient-to-r dark:from-amber-500 dark:to-pink-500 from-pink-500 to-violet-500 bg-clip-text text-transparent inline-block'>
-                &nbsp;things
-              </p>
-              <p className='inline-block motion-preset-seesaw-sm hover:motion-paused motion-preset-oscillate motion-preset-float-sm'>
-                &nbsp;within
-              </p>
-              <p className='inline-block motion-preset-oscillate motion-preset-float-sm hover:motion-paused'>
-                &nbsp;the&nbsp;
-              </p>
-              <p className='inline-block -motion-rotate-loop-5 motion-duration-[4000ms] hover:motion-paused'>
-                realm of
-              </p>
-              <p className='hover:motion-paused motion-preset-seesaw-sm motion-preset-oscillate motion-preset-float-sm bg-gradient-to-r dark:from-pink-500 dark:to-amber-500 from-pink-500 to-violet-500 bg-clip-text text-transparent inline-block'>
-                &nbsp;creativity!&nbsp;
-              </p>
-
-
-              <p className='onest-bold text-lg pt-2 pb-16 pt-10'>
-                {`>`} My greatest passions exist within
-                <p className='bg-gradient-to-r dark:to-indigo-500 dark:via-indigo-400 dark:from-sky-400 to-pink-500 from-violet-500 bg-clip-text text-transparent inline-block'>
-                  &nbsp;software development&nbsp;
-                </p>
-                &
-                <p className='bg-gradient-to-r dark:to-violet-500 dark:from-pink-500 to-pink-500 from-violet-500 bg-clip-text text-transparent inline-block'>
-                  &nbsp;product design&nbsp;
-                </p>
-                <p className='md:pt-0 pt-5'>
-                  {`>`} I also like to spend time on making illustrations and writing prose & poetry.
-                </p>
-              </p>
-            </div>
-
-            <section id="projects" className=''>
-              <div className='flex flex-row justify-end mr-10'>
-                <h1 className='text-6xl onest-bold bg-clip-text text-transparent inline-block bg-gradient-to-r from-violet-600 to-pink-600'>
-                  Projects
-                </h1>
-              </div>
-              <Card />
-              <Card />
-              <Card />
-            </section>
+            <Intro />
+            <Projects />
           </section>
 
           <Footer />
